Unify tabs feature flag
