+++ apps/studio/components/interfaces/App/FeaturePreview/FeaturePreviewContext.tsx
@@ -68,17 +68,15 @@ export const useFeaturePreviewContext = () => useContext(FeaturePreviewContext)
 
 export const FeaturePreviewContextProvider = ({ children }: PropsWithChildren<{}>) => {
   const { hasLoaded } = useContext(FeatureFlagContext)
-
-  const enableSQLEditorTabs = useFlag('featurePreviewSqlEditorTabs')
-  const enableTableEditorTabs = useFlag('featurePreviewTabsTableEditor')
+  const enableTabsInterface = useFlag('tabsInterface')
 
   // [Joshen] Similar logic to feature flagging previews, we can use flags to default opt in previews
   const isDefaultOptIn = (feature: (typeof FEATURE_PREVIEWS)[number]) => {
     switch (feature.key) {
       case LOCAL_STORAGE_KEYS.UI_SQL_EDITOR_TABS:
-        return enableSQLEditorTabs
+        return enableTabsInterface
       case LOCAL_STORAGE_KEYS.UI_TABLE_EDITOR_TABS:
-        return enableTableEditorTabs
+        return enableTabsInterface
       default:
         return false
     }
