+++ apps/studio/components/layouts/Tabs/NewTab.tsx
@@ -1,30 +1,23 @@
 import { PermissionAction } from '@supabase/shared-types/out/constants'
 import { partition } from 'lodash'
-import { ExternalLink, Table2, X } from 'lucide-react'
+import { Table2 } from 'lucide-react'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
 import { toast } from 'sonner'
 
-import { LOCAL_STORAGE_KEYS, useParams } from 'common'
+import { useParams } from 'common'
+import { TabsUpdateCallout } from 'components/interfaces/App/FeaturePreview/TableEditorTabs'
 import { SQL_TEMPLATES } from 'components/interfaces/SQLEditor/SQLEditor.queries'
 import { createSqlSnippetSkeletonV2 } from 'components/interfaces/SQLEditor/SQLEditor.utils'
-import { ButtonTooltip } from 'components/ui/ButtonTooltip'
-import { InlineLinkClassName } from 'components/ui/InlineLink'
 import { useSendEventMutation } from 'data/telemetry/send-event-mutation'
 import { useCheckPermissions } from 'hooks/misc/useCheckPermissions'
-import { useLocalStorageQuery } from 'hooks/misc/useLocalStorage'
 import { useSelectedOrganization } from 'hooks/misc/useSelectedOrganization'
 import { uuidv4 } from 'lib/helpers'
 import { useProfile } from 'lib/profile'
-import { useAppStateSnapshot } from 'state/app-state'
 import { useSqlEditorV2StateSnapshot } from 'state/sql-editor-v2'
 import { useTableEditorStateSnapshot } from 'state/table-editor'
 import { getTabsStore } from 'state/tabs'
 import {
-  Alert_Shadcn_,
-  AlertDescription_Shadcn_,
-  AlertTitle_Shadcn_,
-  Badge,
   Button,
   cn,
   SQL_ICON,
@@ -46,7 +39,6 @@ export function NewTab() {
   const org = useSelectedOrganization()
   const { project } = useProjectContext()
 
-  const appSnap = useAppStateSnapshot()
   const snap = useTableEditorStateSnapshot()
   const snapV2 = useSqlEditorV2StateSnapshot()
 
@@ -54,11 +46,6 @@ export function NewTab() {
   const [quickstarts] = partition(SQL_TEMPLATES, { type: 'quickstart' })
 
   const { mutate: sendEvent } = useSendEventMutation()
-
-  const [tabsInterfaceAcknowledge, setTabsInterfaceAcknowledge] = useLocalStorageQuery(
-    LOCAL_STORAGE_KEYS.TABS_INTERFACE_ACKNOWLEDGED,
-    false
-  )
 
   const canCreateSQLSnippet = useCheckPermissions(PermissionAction.CREATE, 'user_content', {
     resource: { type: 'sql', owner_id: profile?.id },
@@ -129,53 +116,7 @@ export function NewTab() {
   return (
     <div className="bg-surface-100 h-full overflow-y-auto py-12">
       <div className="mx-auto max-w-2xl flex flex-col gap-10 px-10">
-        {!tabsInterfaceAcknowledge && (
-          <Alert_Shadcn_ className="mb-4 relative">
-            <AlertTitle_Shadcn_>
-              <Badge variant="brand" className="mr-2">
-                NEW
-              </Badge>
-              Tabs Interface for Editors
-            </AlertTitle_Shadcn_>
-            <AlertDescription_Shadcn_>
-              The Table and SQL Editors now feature tabs for improved navigation and organization!
-              Check out our{' '}
-              <span
-                className={cn(InlineLinkClassName, 'cursor-pointer')}
-                onClick={() => {
-                  appSnap.setSelectedFeaturePreview(
-                    editor === 'table'
-                      ? LOCAL_STORAGE_KEYS.UI_TABLE_EDITOR_TABS
-                      : LOCAL_STORAGE_KEYS.UI_SQL_EDITOR_TABS
-                  )
-                  appSnap.setShowFeaturePreviewModal(true)
-                }}
-              >
-                feature previews
-              </span>{' '}
-              for more information.
-            </AlertDescription_Shadcn_>
-            <AlertDescription_Shadcn_ className="mt-4 flex items-center gap-x-2">
-              <Button asChild type="default" icon={<ExternalLink />}>
-                <a
-                  target="_blank"
-                  rel="noreferrer noopener"
-                  href="https://github.com/orgs/supabase/discussions/35636"
-                >
-                  View announcement
-                </a>
-              </Button>
-            </AlertDescription_Shadcn_>
-            <ButtonTooltip
-              type="text"
-              icon={<X />}
-              className="absolute top-2 right-2 px-1"
-              onClick={() => setTabsInterfaceAcknowledge(true)}
-              tooltip={{ content: { side: 'bottom', text: 'Dismiss' } }}
-            />
-          </Alert_Shadcn_>
-        )}
-
+        <TabsUpdateCallout />
         <div className="grid grid-cols-2 gap-4">
           {actions.map((item, i) => (
             <ActionCard key={`action-card-${i}`} {...item} />
